+++ src/controllers/api/index.js
@@ -1,0 +1,8 @@
+const { Router } = require('express');
+const productController = require('./product.controller');
+
+const router = new Router();
+
+router.use('/view', productController);
+
+module.exports = router;
